Fix upload error on macosx
The platformio upload command fails on macOS. This is because the
path to artemis_svl determined by the python builder script (main.py)
is incorrect due to a bad conditional assignment. The condition for
setting the appropriate path on a macOS was guaranteed to be always
false because a lower case word is compared to a capitalize.